Closes #36, cleans out unrefrenced local variables. Those, that are used later in commented code or are likely to return, were commented out; others were deleted. This change also addresses some of the type conversion warnings, function declarations, and similar low-hanging fruits for this ticket.
git-svn-id: http://developer.mangband.org/svn/mangband/trunk@712 802d84f3-b497-254e-bede-06c0a5a3456a

+++ src/server/birth.c
@@ -1128,7 +1128,7 @@ static void player_setup(int Ind)
 	{
 		/* Add */
 		add_player_name(p_ptr->name, p_ptr->id);
-	printf("Player Name is [%s], id is %d\n",p_ptr->name, p_ptr->id);
+	printf("Player Name is [%s], id is %d\n", p_ptr->name, (int)p_ptr->id);
 	}
 
 	/* Set his "current activities" variables */
